requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.36
feedparser>=6.0.11
openai>=1.54.0
python-dotenv>=1.0.1
requests>=2.32.3
apscheduler>=3.10.4
pydantic>=2.10.0
beautifulsoup4>=4.12.0
google-generativeai>=0.8.0
anthropic>=0.39.0
